rename notify function and move to new file
